+++ src/main/javascript/AppInfoService.js
@@ -1,3 +1,5 @@
+import {selectDefinitionsFromManifest} from './Setting';
+
 export class AppInfoService
 {
   /**
@@ -19,24 +21,42 @@ export class AppInfoService
    * @param instanceId
    * @return {Promise.<{}>}
    */
-  determineInstallType({ api, instanceId })
+  getInstanceInstallStatus({ api, instanceId })
   {
     return api.get(`apps/${instanceId}/status`)
       .then(({ body }) => {
         return body;
       })
-      .then(({ is_installed }) => {
-        return is_installed ? 'update' : 'first-time';
-      })
-      ;
+      .then(({ is_installed }) => is_installed)
+    ;
+  }
+
+  loadManifestChanges({ api, appId  })
+  {
+    return api.get(`apps/packages/${appId}/manifest-changes`).then(({ body }) => body);
   }
 
   /**
    * @return {Promise.<{}>}
    */
-  loadManifest({ api, instanceId  })
+  loadManifest({ api, appId  })
   {
-    return api.get(`apps/${instanceId}/manifest`).then(({ body }) => body);
+    return api.get(`apps/packages/${appId}/manifest`).then(({ body }) => body);
+  }
+
+  /**
+   * @param {{}} manifest
+   * @param {Function} storage
+   * @return {*}
+   */
+  loadSettingValues(manifest, storage) {
+    const settings = selectDefinitionsFromManifest(manifest);
+    if (settings.length) {
+      const names = settings.map(setting => setting.name);
+      return storage.getAppStorage(names);
+    }
+
+    return Promise.resolve({});
   }
 
 }